requirements.txt
flask==2.3.3
python-dotenv==1.0.0
flask-bcrypt
flask-login
flask_sqlalchemy==3.1.1
flask-migrate
ics>=0.7.2
requests>=2.28.0
icalendar>=5.0.0
pytz==2024.1
SQLAlchemy==2.0.40
psycopg2-binary
gunicorn